Extract "SectionInfo<> into "NodeInfo.h" file
